requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.0
defusedxml==0.7.1
django-allauth==0.52.0
django-autoslug==1.9.8
django-bootstrap==0.2.4
django-crispy-forms==1.13.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.14
whitenoise==5.3.0
wincertstore==0.2
gunicorn==20.0.4
Pillow==9.1.0
psycopg2-binary==2.9.1
